requirements.txt
praw==7.7.1
pandas==2.0.0
scikit-learn==1.2.2
fastapi==0.95.1
uvicorn==0.21.1
python-dotenv==1.0.0
requests==2.28.2
beautifulsoup4==4.12.2
jupyter==1.0.0
mlflow==2.3.0
apache-airflow==2.6.0
